move keyEventData class to the liveblog package
