+++ article/app/views/liveblog/keyEvents.scala.html
@@ -1,4 +1,4 @@
-@import model.KeyEventData
+@import model.liveblog.KeyEventData
 @(id: String, keyEvents: Seq[KeyEventData])
 
 @keyEvents.map { keyEvent =>
