+++ article/app/views/liveblog/liveBlogBody.scala.html
@@ -1,6 +1,6 @@
 @import _root_.liveblog.BodyBlocks
 @import model.liveblog.BodyBlock
-@import model.KeyEventData
+@import model.liveblog.KeyEventData
 @(model: LiveBlogPage, blocks: BodyBlocks[BodyBlock])(implicit request: RequestHeader)
 
 @import common.LinkTo
